@@ -15,35 +15,35 @@
   "dependencies": {
     "@babel/core": "^7.12.3",
     "@babel/preset-react": "^7.12.5",
-    "@types/node": "^16.6.1",
-    "@types/react": "^17.0.17",
-    "@types/react-dom": "^17.0.9",
+    "@types/node": "^18.11",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "aes-js": "^3.1.2",
-    "babel-loader": "^8.2.0",
+    "babel-loader": "^9",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "cross-env": "^7.0.2",
     "css-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "mini-css-extract-plugin": "^2.2.0",
-    "postcss-loader": "^6.1.1",
-    "react": "^17",
-    "react-dom": "^17",
+    "postcss-loader": "^7",
+    "react": "^18",
+    "react-dom": "^18",
     "rimraf": "^3.0.2",
     "style-loader": "^3.2.1",
-    "tailwindcss": "^2.2.7",
+    "tailwindcss": "^3",
     "webpack": "^5.4",
-    "webpack-cli": "^4.2"
+    "webpack-cli": "^5"
   },
   "devDependencies": {
     "@types/aes-js": "^3.1.1",
-    "@typescript-eslint/eslint-plugin": "^4.7",
-    "@typescript-eslint/parser": "^4.7",
+    "@typescript-eslint/eslint-plugin": "^5",
+    "@typescript-eslint/parser": "^5",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.13",
+    "eslint": "^8",
     "eslint-plugin-react": "^7.21",
     "prettier": "^2.1",
     "ts-loader": "^9.2.5",
     "typescript": "^4.0",
-    "webpack-dev-server": "^3.11.0"
+    "webpack-dev-server": "^4"
   }
 }
